fix(ci): change release-type from gradle to java
Release Please v4 doesn't support 'gradle' as a release type. For Java/
Gradle projects, we need to use 'java' as the release-type. This fixes
the error 'Unknown release type: gradle' in the Release Please workflow.

Updated both release-please.yml and release-please-config.json to use
'java' instead of 'gradle' for the release-type configuration.